split shared helper function into beter utils file
