+++ seqr/utils/search/add_data_utils.py
@@ -1,12 +1,24 @@
-from seqr.models import Sample
+from collections import defaultdict, OrderedDict
+from django.contrib.auth.models import User
+from django.db.models import F
+
+from reference_data.models import GENOME_VERSION_GRCh38, GENOME_VERSION_LOOKUP
+from seqr.models import Sample, Individual, Project
 from seqr.utils.communication_utils import send_project_notification, safe_post_to_slack
+from seqr.utils.logging_utils import SeqrLogger
 from seqr.utils.search.utils import backend_specific_call
 from seqr.utils.search.elasticsearch.es_utils import validate_es_index_metadata_and_get_samples
 from seqr.views.utils.airtable_utils import AirtableSession, ANVIL_REQUEST_TRACKING_TABLE
 from seqr.views.utils.dataset_utils import match_and_update_search_samples, load_mapping_file
+from seqr.views.utils.export_utils import write_multiple_files_to_gs
 from seqr.views.utils.permissions_utils import is_internal_anvil_project, project_has_anvil
 from settings import SEQR_SLACK_DATA_ALERTS_NOTIFICATION_CHANNEL, BASE_URL, ANVIL_UI_URL, \
     SEQR_SLACK_ANVIL_DATA_LOADING_CHANNEL
+
+logger = SeqrLogger(__name__)
+
+
+SEQR_V3_PEDIGREE_GS_PATH = 'gs://seqr-loading-temp/v3.1'
 
 
 def _hail_backend_error(*args, **kwargs):
@@ -91,3 +103,60 @@ def notify_search_data_loaded(project, dataset_type, sample_type, inactivated_sa
         subject='New data available in seqr',
     )
     return is_internal
+
+
+def prepare_data_loading_request(projects: list[Project], sample_type: str, dataset_type: str, data_path: str, user: User,
+                                 genome_version: str = GENOME_VERSION_GRCh38, is_internal: bool = False,
+                                 individual_ids: list[str] = None):
+    project_guids = sorted([p.guid for p in projects])
+    variables = {
+        'projects_to_run': project_guids,
+        'callset_path': data_path,
+        'sample_source': 'Broad_Internal' if is_internal else 'AnVIL',
+        'sample_type': sample_type,
+        'dataset_type': _dag_dataset_type(sample_type, dataset_type),
+        'reference_genome': GENOME_VERSION_LOOKUP[genome_version],
+    }
+    upload_info = _upload_data_loading_files(projects, user, genome_version, sample_type, dataset_type, individual_ids)
+    return variables, upload_info
+
+
+def _dag_dataset_type(sample_type: str, dataset_type: str):
+    return 'GCNV' if dataset_type == Sample.DATASET_TYPE_SV_CALLS and sample_type == Sample.SAMPLE_TYPE_WES \
+        else dataset_type
+
+
+def _upload_data_loading_files(projects: list[Project], user: User, genome_version: str, sample_type: str, dataset_type: str,
+                               individual_ids: list[str]):
+    file_annotations = OrderedDict({
+        'Project_GUID': F('family__project__guid'), 'Family_GUID': F('family__guid'),
+        'Family_ID': F('family__family_id'),
+        'Individual_ID': F('individual_id'),
+        'Paternal_ID': F('father__individual_id'), 'Maternal_ID': F('mother__individual_id'), 'Sex': F('sex'),
+    })
+    annotations = {'project': F('family__project__guid'), **file_annotations}
+    individual_filter = {'id__in': individual_ids} if individual_ids else {'family__project__in': projects}
+    data = Individual.objects.filter(**individual_filter).order_by('family_id', 'individual_id').values(
+        **dict(annotations))
+
+    data_by_project = defaultdict(list)
+    for row in data:
+        data_by_project[row.pop('project')].append(row)
+
+    info = []
+    header = list(file_annotations.keys())
+    files = [(f'{project_guid}_pedigree', header, rows) for project_guid, rows in data_by_project.items()]
+    gs_path = _get_gs_pedigree_path(genome_version, sample_type, dataset_type)
+    try:
+        write_multiple_files_to_gs(files, gs_path, user, file_format='tsv')
+    except Exception as e:
+        logger.error(f'Uploading Pedigrees to Google Storage failed. Errors: {e}', user, detail={
+            project: rows for project, _, rows in files
+        })
+    info.append(f'Pedigree files have been uploaded to {gs_path}')
+
+    return info
+
+
+def _get_gs_pedigree_path(genome_version: str, sample_type: str, dataset_type: str):
+    return f'{SEQR_V3_PEDIGREE_GS_PATH}/{GENOME_VERSION_LOOKUP[genome_version]}/{dataset_type}/pedigrees/{sample_type}/'
