+++ seqr/views/apis/anvil_workspace_api.py
@@ -303,8 +303,8 @@ def _trigger_add_workspace_data(project, pedigree_records, user, data_path, samp
         *{user.email}* requested to load {num_updated_individuals} new{reload_summary} {sample_type} samples ({GENOME_VERSION_LOOKUP.get(project.genome_version)}) from AnVIL workspace *{project.workspace_namespace}/{project.workspace_name}* at 
         {data_path} to seqr project <{_get_seqr_project_url(project)}|*{project.name}*> (guid: {project.guid})"""
     trigger_success = trigger_data_loading(
-        [project], sample_type, Sample.DATASET_TYPE_VARIANT_CALLS, data_path, user, success_message,
-        SEQR_SLACK_ANVIL_DATA_LOADING_CHANNEL, f'ERROR triggering AnVIL loading for project {project.guid}',
+        [project], sample_type, Sample.DATASET_TYPE_VARIANT_CALLS, data_path, user=user, success_message=success_message,
+        success_slack_channel=SEQR_SLACK_ANVIL_DATA_LOADING_CHANNEL, error_message=f'ERROR triggering AnVIL loading for project {project.guid}',
         genome_version=project.genome_version,
     )
     AirtableSession(user, base=AirtableSession.ANVIL_BASE).safe_create_records(
